+++ spec/support/pages/overlays/adhoc_overlay.rb
@@ -4,10 +4,26 @@ class AdhocOverlay < CardOverlay
     find('.adhoc-content-toolbar .fa-plus')
   end
 
-  def attach_and_upload_file(file_name)
+  def upload_attachment(file_name)
     add_content_button.click
     upload_file(element_id: "file_attachment",
                 file_name: file_name,
                 sentinel: Proc.new{ Attachment.count })
+  end
+
+  def replace_attachment(file_name)
+    find(".thumbnail-preview").hover
+    find(".replace").click
+
+    within(".replace") do
+      file_input_id = page.find("input[type=file]", visible: false)[:id]
+      upload_file(
+        element_id: file_input_id,
+        file_name: file_name,
+        sentinel: Proc.new {
+          process_sidekiq_jobs
+          Attachment.last[:file] == file_name
+        })
+    end
   end
 end